Move hostname verification temporarily to sub package
