+++ common/federation/src/main/java/app/coronawarn/server/common/federation/client/hostname/HostnameVerifierProvider.java
@@ -1,4 +1,4 @@
-package app.coronawarn.server.common.federation.client;
+package app.coronawarn.server.common.federation.client.hostname;
 
 import javax.net.ssl.HostnameVerifier;
