+++ common/federation/src/test/java/app/coronawarn/server/common/federation/client/CloudFederationFeignHttpClientProviderSmokeTest.java
@@ -4,6 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import app.coronawarn.server.common.federation.client.config.FederationGatewayConfig;
 import app.coronawarn.server.common.federation.client.config.FederationGatewayConfig.Ssl;
+import app.coronawarn.server.common.federation.client.hostname.DefaultHostnameVerifierProvider;
 import java.io.File;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
