+++ services/submission/src/test/java/app/coronawarn/server/services/submission/ServerApplicationTests.java
@@ -3,7 +3,7 @@ package app.coronawarn.server.services.submission;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
-import app.coronawarn.server.common.federation.client.NoopHostnameVerifierProvider;
+import app.coronawarn.server.common.federation.client.hostname.NoopHostnameVerifierProvider;
 import app.coronawarn.server.services.submission.controller.SubmissionController;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
